fix: resolve MSVC warnings for pointer truncation and format mismatches
- memory_pool.h: add parentheses around subtraction before cast to
  (long int) so the full expression is cast, not just the first operand
- service.c: change buffer_size from size_t to DWORD to match
  WSAAddressToString's LPDWORD parameter on Win64
- file.c: replace (DWORD) NULL with 0 for MultiByteToWideChar and
  WideCharToMultiByte dwFlags parameter, avoiding 64-to-32 truncation

+++ src/viriatum/system/service.c
@@ -1573,7 +1573,7 @@ ERROR_CODE resolve_connection(struct connection_t *connection, SOCKET_ADDRESS *s
     that are going to be used in the "custom" parsing of the
     ipv6 addresses under the windows platform */
     size_t index;
-    size_t buffer_size;
+    DWORD buffer_size;
     unsigned char buffer[64];
 #endif
